Use int64 explicitly

+++ 1016/main.go
@@ -6,19 +6,19 @@ import (
 )
 
 func main() {
-	var min, max int
+	var min, max int64
 	fmt.Scan(&min, &max)
 	fmt.Println(solve(min, max))
 }
 
-func solve(min, max int) int {
-	squares := make(map[int]bool)
+func solve(min, max int64) int {
+	squares := make(map[int64]bool)
 
 	for i := 2; i <= int(math.Sqrt(float64(max))); i++ {
-		iSq := i * i
+		iSq := int64(i) * int64(i)
 
-		a := int(math.Ceil(float64(min) / float64(iSq)))
-		b := int(math.Floor(float64(max) / float64(iSq)))
+		a := int64(math.Ceil(float64(min) / float64(iSq)))
+		b := int64(math.Floor(float64(max) / float64(iSq)))
 
 		for j := a; j <= b; j++ {
 			x := iSq * j
@@ -26,6 +26,6 @@ func solve(min, max int) int {
 		}
 	}
 
-	length := max - min
+	length := int(max - min)
 	return length - len(squares) + 1
 }
